+++ fourtoon-cookie/src/components/auth/GoogleSignInAndSignUpButton.tsx
@@ -5,8 +5,8 @@ import {
 import {CLIENT_ID} from '@env'
 import {supabaseSignInAndSignUpWithIdToken} from "../../apis/supabase";
 import type {Session} from "../../types/session";
-import {OAuthProvider} from "../../types/OAuthProvider";
-import {getGoogleIdTokenWithNativeLogin} from "../../apis/googleSignInAndSignUp";
+import { OAuthProvider } from "../../types/OAuthProvider";
+import {getGoogleIdTokenWithNativeLogin} from "../../apis/googleOAuth";
 
 const GoogleSignInAndSignUpButton = () => {
     GoogleSignin.configure({
